Let's see how this goes
I'm sure this will crumble like... something crumble-y as soon as it
hits production data, but yolo

Anyway, this is the first iteration of the round stats viewer. So far,
cargo and mining have their own stats. The round text at the top of the
page is still WIP though. Hooray inconsistent data!