More updates to round stats
I can't stop refactoring the round class. This includes new data about
the game mode, any objectives, the rounded server population, radio
usage and a buncha other information.

There's also a disabled method to tie the given name of a station to
the round in which it occurred, but this still needs to finessing.